+++ terraform/modules/essbase/scripts/wait.sh
@@ -1,6 +1,6 @@
 #!/bin/bash 
 #
-# Copyright (c) 2019, Oracle and/or its affiliates. All rights reserved.
+# Copyright (c) 2020, Oracle and/or its affiliates. All rights reserved.
 #
 
 time=$1