+++ terraform/modules/load-balancer/variables.tf
@@ -2,48 +2,58 @@
 ## All rights reserved. The Universal Permissive License (UPL), Version 1.0 as shown at http://oss.oracle.com/licenses/upl
 
 variable "enabled" {
+  type    = bool
   default = true
 }
 
 # OCI Service
 variable "compartment_id" {
   description = "Compartment OCID where the loadbalancer VCN is created."
+  type        = string
 }
 
-variable "subnet_count" {}
+variable "subnet_count" {
+  type = number
+}
 
 variable "subnet_ids" {
   description = "The subnet id for the Essbase node."
-  type        = "list"
+  type        = list(string)
 }
 
 variable "node_count" {
   description = "The number of nodes in the cluster...workaround for count issue"
+  type        = number
 }
 
 variable "node_ip_addresses" {
   description = "The target ip addresses of the Essbase nodes"
-  type        = "list"
+  type        = list(string)
 }
 
 variable "display_name_prefix" {
   description = "Display name prefix for the resources created."
+  type        = string
 }
 
 variable "shape" {
   description = "Shape of the load balancer"
+  type        = string
   default     = "100Mbps"
 }
 
 variable "idle_timeout" {
+  type    = number
   default = 300
 }
 
-variable "enable_https" {
-  default = true
+// Tags
+variable "freeform_tags" {
+  type = map(string)
+  default = null
 }
 
-variable "demo_ca" {
-  default = {}
-  type    = "map"
+variable "defined_tags" {
+  type = map(string)
+  default = null
 }